# Speed up xor_byte_strings by 70%