@printercu 2018-05-27
It's known that block loops are slightly slower, buts it's idiomatic in ruby. This does not mean that all loops are to be rewritten with `while` loops. Please consider this solution:

```ruby
def xor_byte_strings_new(s1, s2) # :doc:
  s2 = s2.dup
  s1.size.times { |i| s2.setbyte(i, s1.getbyte(i) ^ s2.getbyte(i)) }
  s2
end
```

It's 20% slower comparing to `while` version.