@gregmolnar 2018-05-28
@printercu I think that 20% matters more in this case than keeping the code idiomatic in ruby.