@printercu 2018-05-28
Well, let's start further optimisations from this file. There is `#any?` call: https://github.com/jeremyevans/rails/blob/8b10a9414dd30817b1fc9c4c8cb7600cca0d15b3/actionpack/lib/action_controller/metal/request_forgery_protection.rb#L295
I've just benchmarked, and while version brings 26% performance improvement. It's almost 1.5 times more than 20% in previous optimisation.

```ruby
require 'benchmark/ips'
require 'securerandom'

def with_while(ar) # :doc:
  size = ar.size
  i = 0
  while i < size
    return true if ar[i] > 5
    i += 1
  end
  false
end

def with_block(ar) # :doc:
  ar.any? { |x| x > 5 }
end

ar = 10.times.to_a

Benchmark.ips do |x|
  x.report("while"){with_while(ar)}
  x.report("any"){with_block(ar)}
  x.compare!
end
```

```
Warming up --------------------------------------
               while   138.111k i/100ms
                 any   119.997k i/100ms
Calculating -------------------------------------
               while      2.435M (± 6.0%) i/s -     12.154M in   5.012798s
                 any      1.937M (± 5.6%) i/s -      9.720M in   5.034151s

Comparison:
               while:  2434803.1 i/s
                 any:  1936958.7 i/s - 1.26x  slower
```