@gregmolnar 2018-05-28
@printercu I would go with the fastest implementation which seems to be the while loop, but I am not the one who makes the call here.