@printercu 2018-05-28
I'm really sorry, it was a joke. Ruby is not a right language for micro-optimisations. It always was about readability and programmer's happiness. Please see my example with `#any?` again. There are 7 times more code in while-version, and it's far from as readable as block-version.