@gregmolnar 2018-05-28
I don't agree on that. There are cases where micro-optimisations matter more than readability or the length of the code. If you can make a quite often called chunk of code almost 2x faster while making it less idiomatic, I believe that's the right thing to do.